Added BaseController and SiteView
Chaplin supports use of regions which abstract the process of defining
and referencing DOM elements. For example, instead of adding a view to
the `#main` element, we can add it to the `content` region.

The new BaseController is responsible for rendering a SiteView in its
beforeAction.

The SiteView renders a template that sets up the views for the site and
then registers the `content` region as a reference to `#content`. (Note
that we can change this in the future without having to change the other
controllers.)

This commit also updates the WelcomeController to use the new
BaseController and `content` region.