Add @getTemplateFunction to BaseCollectionView
This makes it easier to render a containing template for a
BaseCollectionView. For example, you may want to have a heading on top
of the list:

h2 This is the heading

ul.the-list

Then in the CollectionView you could do:

class MyCollectionView extends BaseCollectionView

  template: require './template'

  itemView: require './some_item_view'

  listSelector: 'ul.the-list'

The `listSelector` property is necessary to tell the CollectionView
which element should serve as the container for the item views.