Update TodoView on model change - approach 1
In our first approach at updating the TodoView when the model changes
we'll listen for `change` events and then update the view accordingly.
While completely re-rendering the view may be fine for simple cases like
this it becomes a bit heavy-handed when the view represents multiple
attributes that may change

We would rather update only the elements whose model attributes have
changed.