Extracted base views to new module
One main goal of our stack is to be able to reuse core components
without having to rewrite or copy and paste code. In this branch we're
starting to extract a new module called trabian-webapp-core. So far
we've moved `app/views/base` to `core/views/base` and
`app/views/base/collection` to `core/views/base/collection` along with
the associated tests.

Thanks to browserify, all we have to do to make these base views
available to our app is add a new aliasMapping to our Gruntfile.